+++ src/Twilio/Rest/Conversations/V1/Service/ConfigurationList.php
@@ -13,14 +13,18 @@ use Twilio\Exceptions\TwilioException;
 use Twilio\InstanceContext;
 use Twilio\ListResource;
 use Twilio\Rest\Conversations\V1\Service\Configuration\NotificationList;
+use Twilio\Rest\Conversations\V1\Service\Configuration\WebhookList;
 use Twilio\Version;
 
 /**
  * @property NotificationList $notifications
+ * @property WebhookList $webhooks
  * @method \Twilio\Rest\Conversations\V1\Service\Configuration\NotificationContext notifications()
+ * @method \Twilio\Rest\Conversations\V1\Service\Configuration\WebhookContext webhooks()
  */
 class ConfigurationList extends ListResource {
     protected $_notifications = null;
+    protected $_webhooks = null;
 
     /**
      * Construct the ConfigurationList
@@ -44,6 +48,17 @@ class ConfigurationList extends ListResource {
         }
 
         return $this->_notifications;
+    }
+
+    /**
+     * Access the webhooks
+     */
+    protected function getWebhooks(): WebhookList {
+        if (!$this->_webhooks) {
+            $this->_webhooks = new WebhookList($this->version, $this->solution['chatServiceSid']);
+        }
+
+        return $this->_webhooks;
     }
 
     /**