+++ src/Twilio/Rest/FlexApi/V1/ConfigurationInstance.php
@@ -57,6 +57,8 @@ use Twilio\Version;
  * @property array $markdown
  * @property string $url
  * @property array $flexInsightsHr
+ * @property bool $flexInsightsDrilldown
+ * @property string $flexUrl
  */
 class ConfigurationInstance extends InstanceResource {
     /**
@@ -110,6 +112,8 @@ class ConfigurationInstance extends InstanceResource {
             'markdown' => Values::array_get($payload, 'markdown'),
             'url' => Values::array_get($payload, 'url'),
             'flexInsightsHr' => Values::array_get($payload, 'flex_insights_hr'),
+            'flexInsightsDrilldown' => Values::array_get($payload, 'flex_insights_drilldown'),
+            'flexUrl' => Values::array_get($payload, 'flex_url'),
         ];
 
         $this->solution = [];