+++ src/Twilio/Rest/Supersim/V1.php
@@ -14,6 +14,7 @@ use Twilio\Exceptions\TwilioException;
 use Twilio\InstanceContext;
 use Twilio\Rest\Supersim\V1\CommandList;
 use Twilio\Rest\Supersim\V1\FleetList;
+use Twilio\Rest\Supersim\V1\IpCommandList;
 use Twilio\Rest\Supersim\V1\NetworkAccessProfileList;
 use Twilio\Rest\Supersim\V1\NetworkList;
 use Twilio\Rest\Supersim\V1\SimList;
@@ -24,6 +25,7 @@ use Twilio\Version;
 /**
  * @property CommandList $commands
  * @property FleetList $fleets
+ * @property IpCommandList $ipCommands
  * @property NetworkList $networks
  * @property NetworkAccessProfileList $networkAccessProfiles
  * @property SimList $sims
@@ -31,6 +33,7 @@ use Twilio\Version;
  * @property UsageRecordList $usageRecords
  * @method \Twilio\Rest\Supersim\V1\CommandContext commands(string $sid)
  * @method \Twilio\Rest\Supersim\V1\FleetContext fleets(string $sid)
+ * @method \Twilio\Rest\Supersim\V1\IpCommandContext ipCommands(string $sid)
  * @method \Twilio\Rest\Supersim\V1\NetworkContext networks(string $sid)
  * @method \Twilio\Rest\Supersim\V1\NetworkAccessProfileContext networkAccessProfiles(string $sid)
  * @method \Twilio\Rest\Supersim\V1\SimContext sims(string $sid)
@@ -39,6 +42,7 @@ use Twilio\Version;
 class V1 extends Version {
     protected $_commands;
     protected $_fleets;
+    protected $_ipCommands;
     protected $_networks;
     protected $_networkAccessProfiles;
     protected $_sims;
@@ -67,6 +71,13 @@ class V1 extends Version {
             $this->_fleets = new FleetList($this);
         }
         return $this->_fleets;
+    }
+
+    protected function getIpCommands(): IpCommandList {
+        if (!$this->_ipCommands) {
+            $this->_ipCommands = new IpCommandList($this);
+        }
+        return $this->_ipCommands;
     }
 
     protected function getNetworks(): NetworkList {