+++ src/Twilio/Rest/Supersim/V1/IpCommandList.php
@@ -7,19 +7,21 @@
  * /       /
  */
 
-namespace Twilio\Rest\Video\V1;
+namespace Twilio\Rest\Supersim\V1;
 
 use Twilio\Exceptions\TwilioException;
 use Twilio\ListResource;
 use Twilio\Options;
-use Twilio\Serialize;
 use Twilio\Stream;
 use Twilio\Values;
 use Twilio\Version;
 
-class RoomList extends ListResource {
+/**
+ * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
+ */
+class IpCommandList extends ListResource {
     /**
-     * Construct the RoomList
+     * Construct the IpCommandList
      *
      * @param Version $version Version that contains the resource
      */
@@ -29,41 +31,39 @@ class RoomList extends ListResource {
         // Path Solution
         $this->solution = [];
 
-        $this->uri = '/Rooms';
+        $this->uri = '/IpCommands';
     }
 
     /**
-     * Create the RoomInstance
+     * Create the IpCommandInstance
      *
+     * @param string $sim The sid or unique_name of the Super SIM to send the IP
+     *                    Command to
+     * @param string $payload The payload to be delivered to the device
+     * @param int $devicePort The device port to which the IP Command will be sent
      * @param array|Options $options Optional Arguments
-     * @return RoomInstance Created RoomInstance
+     * @return IpCommandInstance Created IpCommandInstance
      * @throws TwilioException When an HTTP error occurs.
      */
-    public function create(array $options = []): RoomInstance {
+    public function create(string $sim, string $payload, int $devicePort, array $options = []): IpCommandInstance {
         $options = new Values($options);
 
         $data = Values::of([
-            'EnableTurn' => Serialize::booleanToString($options['enableTurn']),
-            'Type' => $options['type'],
-            'UniqueName' => $options['uniqueName'],
-            'StatusCallback' => $options['statusCallback'],
-            'StatusCallbackMethod' => $options['statusCallbackMethod'],
-            'MaxParticipants' => $options['maxParticipants'],
-            'RecordParticipantsOnConnect' => Serialize::booleanToString($options['recordParticipantsOnConnect']),
-            'VideoCodecs' => Serialize::map($options['videoCodecs'], function($e) { return $e; }),
-            'MediaRegion' => $options['mediaRegion'],
-            'RecordingRules' => Serialize::jsonObject($options['recordingRules']),
-            'AudioOnly' => Serialize::booleanToString($options['audioOnly']),
-            'MaxParticipantDuration' => $options['maxParticipantDuration'],
+            'Sim' => $sim,
+            'Payload' => $payload,
+            'DevicePort' => $devicePort,
+            'PayloadType' => $options['payloadType'],
+            'CallbackUrl' => $options['callbackUrl'],
+            'CallbackMethod' => $options['callbackMethod'],
         ]);
 
         $payload = $this->version->create('POST', $this->uri, [], $data);
 
-        return new RoomInstance($this->version, $payload);
+        return new IpCommandInstance($this->version, $payload);
     }
 
     /**
-     * Streams RoomInstance records from the API as a generator stream.
+     * Streams IpCommandInstance records from the API as a generator stream.
      * This operation lazily loads records as efficiently as possible until the
      * limit
      * is reached.
@@ -90,7 +90,7 @@ class RoomList extends ListResource {
     }
 
     /**
-     * Reads RoomInstance records from the API as a list.
+     * Reads IpCommandInstance records from the API as a list.
      * Unlike stream(), this operation is eager and will load `limit` records into
      * memory before returning.
      *
@@ -103,30 +103,30 @@ class RoomList extends ListResource {
      *                        page_size is defined but a limit is defined, read()
      *                        will attempt to read the limit with the most
      *                        efficient page size, i.e. min(limit, 1000)
-     * @return RoomInstance[] Array of results
+     * @return IpCommandInstance[] Array of results
      */
     public function read(array $options = [], int $limit = null, $pageSize = null): array {
         return \iterator_to_array($this->stream($options, $limit, $pageSize), false);
     }
 
     /**
-     * Retrieve a single page of RoomInstance records from the API.
+     * Retrieve a single page of IpCommandInstance records from the API.
      * Request is executed immediately
      *
      * @param array|Options $options Optional Arguments
      * @param mixed $pageSize Number of records to return, defaults to 50
      * @param string $pageToken PageToken provided by the API
      * @param mixed $pageNumber Page Number, this value is simply for client state
-     * @return RoomPage Page of RoomInstance
+     * @return IpCommandPage Page of IpCommandInstance
      */
-    public function page(array $options = [], $pageSize = Values::NONE, string $pageToken = Values::NONE, $pageNumber = Values::NONE): RoomPage {
+    public function page(array $options = [], $pageSize = Values::NONE, string $pageToken = Values::NONE, $pageNumber = Values::NONE): IpCommandPage {
         $options = new Values($options);
 
         $params = Values::of([
+            'Sim' => $options['sim'],
+            'SimIccid' => $options['simIccid'],
             'Status' => $options['status'],
-            'UniqueName' => $options['uniqueName'],
-            'DateCreatedAfter' => Serialize::iso8601DateTime($options['dateCreatedAfter']),
-            'DateCreatedBefore' => Serialize::iso8601DateTime($options['dateCreatedBefore']),
+            'Direction' => $options['direction'],
             'PageToken' => $pageToken,
             'Page' => $pageNumber,
             'PageSize' => $pageSize,
@@ -134,32 +134,32 @@ class RoomList extends ListResource {
 
         $response = $this->version->page('GET', $this->uri, $params);
 
-        return new RoomPage($this->version, $response, $this->solution);
+        return new IpCommandPage($this->version, $response, $this->solution);
     }
 
     /**
-     * Retrieve a specific page of RoomInstance records from the API.
+     * Retrieve a specific page of IpCommandInstance records from the API.
      * Request is executed immediately
      *
      * @param string $targetUrl API-generated URL for the requested results page
-     * @return RoomPage Page of RoomInstance
+     * @return IpCommandPage Page of IpCommandInstance
      */
-    public function getPage(string $targetUrl): RoomPage {
+    public function getPage(string $targetUrl): IpCommandPage {
         $response = $this->version->getDomain()->getClient()->request(
             'GET',
             $targetUrl
         );
 
-        return new RoomPage($this->version, $response, $this->solution);
+        return new IpCommandPage($this->version, $response, $this->solution);
     }
 
     /**
-     * Constructs a RoomContext
+     * Constructs a IpCommandContext
      *
      * @param string $sid The SID that identifies the resource to fetch
      */
-    public function getContext(string $sid): RoomContext {
-        return new RoomContext($this->version, $sid);
+    public function getContext(string $sid): IpCommandContext {
+        return new IpCommandContext($this->version, $sid);
     }
 
     /**
@@ -168,6 +168,6 @@ class RoomList extends ListResource {
      * @return string Machine friendly representation
      */
     public function __toString(): string {
-        return '[Twilio.Video.V1.RoomList]';
+        return '[Twilio.Supersim.V1.IpCommandList]';
     }
 }