+++ src/Twilio/Rest/Video/V1/RoomInstance.php
@@ -32,6 +32,7 @@ use Twilio\Version;
  * @property int $duration
  * @property string $type
  * @property int $maxParticipants
+ * @property int $maxParticipantDuration
  * @property int $maxConcurrentPublishedTracks
  * @property bool $recordParticipantsOnConnect
  * @property string[] $videoCodecs
@@ -70,6 +71,7 @@ class RoomInstance extends InstanceResource {
             'duration' => Values::array_get($payload, 'duration'),
             'type' => Values::array_get($payload, 'type'),
             'maxParticipants' => Values::array_get($payload, 'max_participants'),
+            'maxParticipantDuration' => Values::array_get($payload, 'max_participant_duration'),
             'maxConcurrentPublishedTracks' => Values::array_get($payload, 'max_concurrent_published_tracks'),
             'recordParticipantsOnConnect' => Values::array_get($payload, 'record_participants_on_connect'),
             'videoCodecs' => Values::array_get($payload, 'video_codecs'),