+++ src/Twilio/Rest/Video/V1/RoomOptions.php
@@ -32,10 +32,12 @@ abstract class RoomOptions {
      * @param array $recordingRules A collection of Recording Rules
      * @param bool $audioOnly Indicates whether the room will only contain audio
      *                        track participants for group rooms.
+     * @param int $maxParticipantDuration The maximum number of seconds a
+     *                                    Participant can be connected to the room
      * @return CreateRoomOptions Options builder
      */
-    public static function create(bool $enableTurn = Values::NONE, string $type = Values::NONE, string $uniqueName = Values::NONE, string $statusCallback = Values::NONE, string $statusCallbackMethod = Values::NONE, int $maxParticipants = Values::NONE, bool $recordParticipantsOnConnect = Values::NONE, array $videoCodecs = Values::ARRAY_NONE, string $mediaRegion = Values::NONE, array $recordingRules = Values::ARRAY_NONE, bool $audioOnly = Values::NONE): CreateRoomOptions {
-        return new CreateRoomOptions($enableTurn, $type, $uniqueName, $statusCallback, $statusCallbackMethod, $maxParticipants, $recordParticipantsOnConnect, $videoCodecs, $mediaRegion, $recordingRules, $audioOnly);
+    public static function create(bool $enableTurn = Values::NONE, string $type = Values::NONE, string $uniqueName = Values::NONE, string $statusCallback = Values::NONE, string $statusCallbackMethod = Values::NONE, int $maxParticipants = Values::NONE, bool $recordParticipantsOnConnect = Values::NONE, array $videoCodecs = Values::ARRAY_NONE, string $mediaRegion = Values::NONE, array $recordingRules = Values::ARRAY_NONE, bool $audioOnly = Values::NONE, int $maxParticipantDuration = Values::NONE): CreateRoomOptions {
+        return new CreateRoomOptions($enableTurn, $type, $uniqueName, $statusCallback, $statusCallbackMethod, $maxParticipants, $recordParticipantsOnConnect, $videoCodecs, $mediaRegion, $recordingRules, $audioOnly, $maxParticipantDuration);
     }
 
     /**
@@ -72,8 +74,10 @@ class CreateRoomOptions extends Options {
      * @param array $recordingRules A collection of Recording Rules
      * @param bool $audioOnly Indicates whether the room will only contain audio
      *                        track participants for group rooms.
+     * @param int $maxParticipantDuration The maximum number of seconds a
+     *                                    Participant can be connected to the room
      */
-    public function __construct(bool $enableTurn = Values::NONE, string $type = Values::NONE, string $uniqueName = Values::NONE, string $statusCallback = Values::NONE, string $statusCallbackMethod = Values::NONE, int $maxParticipants = Values::NONE, bool $recordParticipantsOnConnect = Values::NONE, array $videoCodecs = Values::ARRAY_NONE, string $mediaRegion = Values::NONE, array $recordingRules = Values::ARRAY_NONE, bool $audioOnly = Values::NONE) {
+    public function __construct(bool $enableTurn = Values::NONE, string $type = Values::NONE, string $uniqueName = Values::NONE, string $statusCallback = Values::NONE, string $statusCallbackMethod = Values::NONE, int $maxParticipants = Values::NONE, bool $recordParticipantsOnConnect = Values::NONE, array $videoCodecs = Values::ARRAY_NONE, string $mediaRegion = Values::NONE, array $recordingRules = Values::ARRAY_NONE, bool $audioOnly = Values::NONE, int $maxParticipantDuration = Values::NONE) {
         $this->options['enableTurn'] = $enableTurn;
         $this->options['type'] = $type;
         $this->options['uniqueName'] = $uniqueName;
@@ -85,6 +89,7 @@ class CreateRoomOptions extends Options {
         $this->options['mediaRegion'] = $mediaRegion;
         $this->options['recordingRules'] = $recordingRules;
         $this->options['audioOnly'] = $audioOnly;
+        $this->options['maxParticipantDuration'] = $maxParticipantDuration;
     }
 
     /**
@@ -212,6 +217,18 @@ class CreateRoomOptions extends Options {
      */
     public function setAudioOnly(bool $audioOnly): self {
         $this->options['audioOnly'] = $audioOnly;
+        return $this;
+    }
+
+    /**
+     * The maximum number of seconds a Participant can be connected to the room. The maximum possible value is 86400 seconds (24 hours). The default is 14400 seconds (4 hours).
+     *
+     * @param int $maxParticipantDuration The maximum number of seconds a
+     *                                    Participant can be connected to the room
+     * @return $this Fluent Builder
+     */
+    public function setMaxParticipantDuration(int $maxParticipantDuration): self {
+        $this->options['maxParticipantDuration'] = $maxParticipantDuration;
         return $this;
     }